+++ R/Familiar.R
@@ -6,6 +6,7 @@
 #' @name familiar
 #' @import data.table
 #' @import methods
+#' @importFrom isotree isolation.forest
 #' @importFrom stats predict coef
 #' @importFrom survival Surv coxph survreg
 #' @importFrom utils head tail getFromNamespace